+++ demos/storybook/src/stories/playground/empty-state.stories.ts
@@ -4,10 +4,9 @@ import { EmptyStateModule } from '@pxblue/angular-components';
 import { action } from '@storybook/addon-actions';
 import { text, withKnobs } from '@storybook/addon-knobs';
 import { moduleMetadata, storiesOf } from '@storybook/angular';
-import { UtilModule, wrap } from '../utils';
-const markdown = require('../../../../docs/EmptyState.md');
+import { UtilModule, wrap } from '../../utils';
 
-storiesOf('components/Empty State', module)
+storiesOf('playground/Empty State', module)
     .addDecorator(
         moduleMetadata({
             imports: [EmptyStateModule, MatButtonModule, MatIconModule, UtilModule],
@@ -15,12 +14,6 @@ storiesOf('components/Empty State', module)
     )
     .addDecorator(withKnobs)
     .addDecorator(wrap())
-    .addParameters({
-        notes: { markdown },
-    })
-    .add('Documentation', () => ({
-        template: `<documentation></documentation>`,
-    }))
     .add('with action', () => ({
         template: `
           <pxb-empty-state [title]="title">
